This is partial progress toward abstract vars.  This is a branch commit.
I am giving up on this for a moment.  Unfortunately, string
hacking on variable names is everywhere, not cleanly isolated,
and not clearly specified.  This makes a port to abstract
var names either 1) hard, because all this code has to be cleaned
up, or 2) nonsense, because of the need to covert back and forth
between strings all the time.

I still think it is a good idea, but it will take a fair amount of
effort.