The new LALR changes look good.
Grammar compilation is now about 25 times faster than it
used to be.  There is still room for improvement,
especially in the widespread use of Pervasives.compare,
which is being used in the comparison of production sets,
which are essentially shape list sets.

However, we are in the sub-second range for the grammars
that we have, and this is just a compile-time cost.  Still
we want to keep compile times down.