Made some changes to the Map class.
The keys are now "simple" values, not just strings.  These
include integers, floats, strings, arrays of simple values,
files, and directories.

Only the Map class has the map functions defined.  If you
want to have, say, a File that also includes a Map,
create a subclass that extends both File and Map.

The literal keys now have the form $|key...|.  This works
both for definitions and uses.  The usual modifiers are
allowed $,|key| and $`|key|.

   X. =
      extends $(Map)
      $|key 1| = 1
      $|key 2| = $|key 1| boo
      $|key 2| += foo