Working on outstanding issues for Win32.
This is a fix to be more careful about spaces in filenames,
so that placing ocaml in

   C:\Program Files\Objective Caml

actually works.  This adds the $(quote-argv ...) function that
quotes an array in a way that the command-line parser can recover
the argv string.  This allows include directories with spaces in
a -pp option.