Solved two marshaling problems.
The Mount.t record had functions in it; moved them from venv_inner
to venv_globals.

The primitive functions are also represented by functions.  They could
also be moved to a table in venv_globals, but that would be slow.
Instead, they are function options, and the option is cleared before
marshaling.  In the case where a primitive is unmarshaled, the function
option is restored when it is used.

There was a bug--these lazy primitives need to be remembered in the
list of primitives so they can be cleared when re-marshaled.